V2, troca de alguns nomes e update no readme

+++ Conversor/Invertendo.cpp
@@ -13,7 +13,7 @@ int main()
 string invertendo_string(string frase) {
     string teste2{};
     size_t tamanho_string{ size(frase) };
-    for (int i = tamanho_string - 1;tamanho_string > 0; --i) {
+    for (size_t i = tamanho_string - 1;tamanho_string > 0; --i) {
         tamanho_string -= 1;
         teste2 += frase[i];
